refactor: test helper
