+++ src/test/kotlin/blackjack/card/BlackjackCardsTest.kt
@@ -1,9 +1,7 @@
 package blackjack.card
 
-import domain.card.BlackjackCard
+import blackjack.card.helper.BlackjackCardsTestHelper
 import domain.card.BlackjackCards
-import domain.card.CardNumber
-import domain.card.Suit
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
@@ -25,102 +23,10 @@ class BlackjackCardsTest {
 
     companion object {
         @JvmStatic
-        fun getCardNumberSumLessThanEquals21(): List<Arguments> = listOf(
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.TWO),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.THREE),
-                    ),
-                ),
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.KING),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.NINE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                    ),
-                ),
-            ),
-        )
+        fun getCardNumberSumLessThanEquals21(): List<Arguments> =
+            BlackjackCardsTestHelper.getCardNumberSumLessThanEquals21()
 
         @JvmStatic
-        fun getCardNumberSum(): List<Arguments> = listOf(
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.TWO),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.THREE),
-                    ),
-                ),
-                16,
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.KING),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.NINE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                    ),
-                ),
-                20,
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.KING),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                    ),
-                ),
-                21,
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.DIAMOND, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.FIVE),
-                    ),
-                ),
-                12,
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.DIAMOND, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.FIVE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.FIVE),
-                    ),
-                ),
-                16,
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.TWO),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.THREE),
-                        BlackjackCard(suit = Suit.DIAMOND, number = CardNumber.FOUR),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.FIVE),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.SIX),
-                    ),
-                ),
-                22,
-            ),
-        )
+        fun getCardNumberSum(): List<Arguments> = BlackjackCardsTestHelper.getCardNumberSum()
     }
 }
