+++ src/test/kotlin/blackjack/card/HitTest.kt
@@ -1,9 +1,8 @@
 package blackjack.card
 
+import blackjack.card.helper.HitTestHelper
 import domain.card.BlackjackCard
 import domain.card.BlackjackCards
-import domain.card.CardNumber
-import domain.card.Suit
 import domain.state.Burst
 import domain.state.Hit
 import domain.state.Stand
@@ -52,92 +51,12 @@ class HitTest {
 
     companion object {
         @JvmStatic
-        fun getHitStateTestData(): List<Arguments> = listOf(
-            Arguments.of(
-                BlackjackCards(listOf(BlackjackCard(suit = Suit.CLUB, number = CardNumber.KING))),
-                BlackjackCard(suit = Suit.SPADE, number = CardNumber.FIVE),
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.TWO),
-                    ),
-                ),
-                BlackjackCard(suit = Suit.SPADE, number = CardNumber.FIVE),
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.TEN),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.TEN),
-                    ),
-                ),
-                BlackjackCard(suit = Suit.SPADE, number = CardNumber.ACE),
-            ),
-        )
+        fun getHitStateTestData(): List<Arguments> = HitTestHelper.getHitStateTestData()
 
         @JvmStatic
-        fun getBurstStateTestData(): List<Arguments> = listOf(
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.KING),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.KING),
-                    ),
-                ),
-                BlackjackCard(suit = Suit.SPADE, number = CardNumber.FIVE),
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.TWO),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.SEVEN),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.JACK),
-                    ),
-                ),
-                BlackjackCard(suit = Suit.SPADE, number = CardNumber.TWO),
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.TEN),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.TEN),
-                    ),
-                ),
-                BlackjackCard(suit = Suit.SPADE, number = CardNumber.TWO),
-            ),
-        )
+        fun getBurstStateTestData(): List<Arguments> = HitTestHelper.getBurstStateTestData()
 
         @JvmStatic
-        fun getStandStateTestData(): List<Arguments> = listOf(
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.KING),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.KING),
-                    ),
-                ),
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.ACE),
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.TWO),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.SEVEN),
-                        BlackjackCard(suit = Suit.SPADE, number = CardNumber.JACK),
-                    ),
-                ),
-            ),
-            Arguments.of(
-                BlackjackCards(
-                    listOf(
-                        BlackjackCard(suit = Suit.CLUB, number = CardNumber.TEN),
-                        BlackjackCard(suit = Suit.HEART, number = CardNumber.TEN),
-                    ),
-                ),
-            ),
-        )
+        fun getStandStateTestData(): List<Arguments> = HitTestHelper.getStandStateTestData()
     }
 }
